Redid a bugfix that was reverted in a previous changelist
This was reverted from what I did for some reason. When developer mode is used, I get the local position of the camera and adjust its height to be at the correct height. I was adjust the position instead of the localPosition which was causing the camera to move to the (0,1.5,0) world position instead of local position. This needs to be local position.